requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-crispy-forms==1.9.1
idna==2.8
jsonfield==3.1.0
numpy==1.19.0
oauthlib==3.0.1
pandas==1.0.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
sqlparse==0.3.0
stripe==2.50.0
urllib3==1.24.2
xlrd==1.2.0

